refactor(a2-1099): site address
