+++ packages/forms-web-app/src/journeys/question-overrides/site-address.js
@@ -1,0 +1,75 @@
+/**
+ * @typedef {import('../../dynamic-forms/journey-response').JourneyResponse} JourneyResponse
+ * @typedef {import('../../dynamic-forms/journey').Journey} Journey
+ * @typedef {import('../../dynamic-forms/section').Section} Section
+ * @typedef {import('../../dynamic-forms/question').QuestionViewModel} QuestionViewModel
+ * @typedef {import('../../dynamic-forms/dynamic-components/site-address/question')} SiteAddressQuestion
+ * @typedef {import('appeals-service-api').Api.SubmissionAddress} SubmissionAddress
+ */
+
+/**
+ * Save the answer to the question
+ * @this {SiteAddressQuestion}
+ * @param {import('express').Request} req
+ * @param {import('express').Response} res
+ * @param {Journey} journey
+ * @param {Section} section
+ * @param {JourneyResponse} journeyResponse
+ * @returns {Promise<void>}
+ */
+async function saveAction(req, res, journey, section, journeyResponse) {
+	// check for validation errors
+	const errorViewModel = this.checkForValidationErrors(req, section, journey);
+	if (errorViewModel) {
+		errorViewModel.question = {
+			...errorViewModel.question,
+			value: {
+				addressLine1: req.body[this.fieldName + '_addressLine1'],
+				addressLine2: req.body[this.fieldName + '_addressLine2'],
+				townCity: req.body[this.fieldName + '_townCity'],
+				county: req.body[this.fieldName + '_county'],
+				postcode: req.body[this.fieldName + '_postcode']
+			}
+		};
+
+		return this.renderAction(res, errorViewModel);
+	}
+
+	// save
+	const { address, siteAddressSet, fieldName, addressId } = await this.getDataToSave(
+		req,
+		journeyResponse
+	);
+
+	await req.appealsApiClient.postSubmissionAddress(
+		journeyResponse.journeyId,
+		journeyResponse.referenceId,
+		{
+			...address,
+			fieldName,
+			// @ts-ignore
+			// TODO: check how the API handles getting no id
+			id: addressId
+		}
+	);
+
+	if (siteAddressSet) {
+		await req.appealsApiClient.updateAppellantSubmission(journeyResponse.referenceId, {
+			siteAddress: siteAddressSet
+		});
+		journeyResponse.answers.siteAddress = siteAddressSet;
+	}
+
+	// check for saving errors
+	const saveViewModel = this.checkForSavingErrors(req, section, journey);
+	if (saveViewModel) {
+		return this.renderAction(res, saveViewModel);
+	}
+
+	// move to the next question
+	return this.handleNextQuestion(res, journey, section.segment, this.fieldName);
+}
+
+module.exports = {
+	saveAction
+};
